finish tasklist

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Models\Task;
 
+
 class TasksController extends Controller
 {
     /**
@@ -15,11 +16,25 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
         
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);
+        if (\Auth::check()) { // 認証済みの場合
+            
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->get();
+
+            // indexビューでそれらを表示
+            return view('tasks.index', [
+                    'tasks' => $tasks,
+            ]);
+            
+        }else {
+            
+            return view('dashboard');
+            
+        }
+        
     }
 
     /**
@@ -50,11 +65,11 @@ class TasksController extends Controller
             'content' => 'required',
         ]);
         
-        // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -69,12 +84,23 @@ class TasksController extends Controller
     public function show($id)
     {
         // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            $task = Task::findOrFail($id);
+    
+            if (\Auth::id() === $task->user_id) {
+            
+                // メッセージ詳細ビューでそれを表示
+                return view('tasks.show', [
+                    'task' => $task,
+                ]);
+            
+            }else {
+                
+                // トップページへリダイレクトさせる
+                return redirect('/');
+                
+            }
+            
     }
 
     /**
@@ -87,11 +113,20 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        
+        if (\Auth::id() === $task->user_id) {
+            
+            // メッセージ編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+            
+        }else {
+                
+                // トップページへリダイレクトさせる
+                return redirect('/');
+                
+        }
     }
 
     /**
@@ -110,13 +145,20 @@ class TasksController extends Controller
         
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        
+        if (\Auth::id() === $task->user_id) {
+        
+            // メッセージを更新
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+    
+            // トップページへリダイレクトさせる
+            return redirect('/');
+            
+        }
+        
+        
     }
 
     /**
@@ -129,10 +171,15 @@ class TasksController extends Controller
     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+            // トップページへリダイレクトさせる
+            return redirect('/');
+            
+        }
+
+        
     }
 }
